ai-service/requirements.txt
flask==2.3.3
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
kafka-python==2.0.2
python-dotenv==1.0.0
joblib==1.3.2
gunicorn==21.2.0
flask-cors==4.0.0
prometheus-client==0.17.1
torch
dvc
mlflow
zenml
optuna
